Final implementation of Quicksolve
Committing the begin of the final implementation phase of quicksolve
with the delayed/asynchronous evaluation framework (AEF). The code is
sketched up and untested and, so far, lacks refcounting on the operands
so they are removed once no longer needed. It may be worth noting that
baked expressions carry a bit more of necessary information, notably
"queue", "n_expression_deps", "expression_deps", "dependee_count" and
even "refcount" which are not needed for unbaked expressions ("refcount"
if we strictly forbid redundant evaluations as we currently do).
This may be streamlined by introducing two different types where the one
for the baked expression is derived from the ordinary expression.