Termination of operands discarded reference
The API for qs_operand_terminate was misleading in the sense that one
would loose a reference when discarding the result. With this fix, the
return value is immediately stored back into the coefficients. Two
alternative, probably nicer solutions would be possible:

1) qs_operand_terminate does not actually base a new QsTerminal on the
QsOperand but rather modifies the given QsOperand in-place (keeping its
refcount)

2) qs_operand_terminate does not automatically consume the given
QsOperand. In that case, we would have to discard either of the
references afterwards.

In that sense, the first solution seems to be more convenient and more
work to implement while the second seems simpler and cleaner. The second
solution should probably be implemented in one of the next commits.