Fixes for the waiter
A very nasty bug appeared in the waiter. The origin of the bug was the
attempt to identify two QsTerminals (namely the one having finished
calculation and the one "being currently waited for" as stored in the
AEF) by pointer. The consequence of this was that, since we have
multiple threads, the QsTerminal finished evaluation and before the
worker even came to sending out the signal, the main thread handled the
result, freed the QsTerminal, allocated a new one at the same address
and indicated it would wait for the latter. Then, the worker arrived at
the point where it would signal completion of the old QsTerminal,
founding the pointer in the AEF to seemingly indicate that the AEF was
waiting for it. This caused a premature signaling of the waiting thread,
which went on wrongly.

The waiting mechanism was now cleanly cut out of the AEF and works
differently now, which, however, required another pointer in
BakedExpression.