Load- and Save infrastructure
The foundation for suspending at and restoring from steps during the
process was laid out. This will allow to continue calculations which
were started with IdSolver as well as interrupt the calculation and
continue at a later time.
Also, storing of identities will become important once the memory
management in the pivot graph is implemented to store unused identities
back to disk in order to keeep memory usage low.