Memory usage management
The pivot graph now keeps track of the order in which the pivots were
accessed. This serves as an ad-hoc metric for which pivots (namely those
which haven't been accessed for the longest time) should be written
back.
However, the question which pivots should be written back and when is a
complicated one and this is only a very naive solution. More factors
which could be taken into account but are close to impossible to account
for without a significant overhead are related to

1) Writes to the database blocking reads from the database and we may
thus want to write to databases which are unlikely to be read from

2) Readiness of the pivot's operands (since we can only write fully
terminated operands)

3) Probability of reading a pivot (which is, in a simple fashion, the
problem being addressed by the current metric)

The USAGE_MARGIN constant is a policy dependend value which, as
described, guarantees that at least USAGE_MARGIN-many pivots which were
last explicitly requested are kept in memory.