Memory management and storage
Synchronous memory management is finished. Benchmarks should reveal
whether

a) The heuristics for preventing blocks is sufficient or a more
elaborate prediction scheme must be developed

b) Writing back synchronously (from the vantage point of pivot graph,
though the implementation of IntegralMgr may choose to perform the
writing without blocking) is sufficient. In a more advanced
implementation, the writing back could be performed asynchronously and
lag behind up to a certain amount of integrals.

c) Purging the pivots from the pivot graph is enough to free the memory
or the databases may have to be closed/flushed in order for the memory
to actually be sufficiently freed.

On top of that, termination by SIGINT was implemented in two termination
points in the czakonprime.c policy.