Termination and continuation
Termination is now respected by the loop in main. Self-recursion
prevention, which was previously done based upon the order of pivots
is now more sensibly done by comparing the QsComponent. This is
necessary if continuing from IdSolver solution databases were no order
is stored and the order of the loaded pivot becomes 0.