Recursion by despair
The previous recursion which was set to either "full" or not "full" and
differentiated pivots between "solving" and not "solving" did not work
in all cases and led to infinite recursions.
A finer system which works based upon a measure of "despair" (a higher
value indicating repeated attempts normalize a pivot) and number of
"considerations" (a higher value indicating an increasing amount of
substitutions of that line in the current attempt) now tries to prevent
short-looped infinite recursion.