Almost correct implementation of multiple wait
There is currently a memory leak if an operand has its waiter
overwritten by a new waiter. This will be fixed since in any case we
have to make sure to already check for ready operands while looking for
new ones in order not to load unnecessary data from database.