Reassert load before termination
The workflow of the solution assumed that when czakon_prime would return
from a call on a specific QsComponent, then that QsComponent would be
inside USAGE_MARGIN because appropriately loaded by the invocation. On
termination, however, the termination occurred before the load such that
this assumption no longer held. The load which was assumed was moved
accordingly to now always take place.