Dual pass preparation
The ADC is augmented by a DC count because we may not set change
QsTerminal to be results before all ADC decrements have been performed,
because decrementing them may recollecting them which in turn will need
access to all QsTerminal which have decrements pending.

The current approach, in comparison with the original implementation, is
heavy locked/mutexed and by a factor 10 slower than the original
implementation in the regime of small coefficients, also owing to the
lack of removing zeros in the policy (which will be adressed in the
coming commits).