Various fixes and new flag
A new flag was added to control how the policy deals with edges which
are numerically zero, whether they are to be discarded, kept, or waited
upon. The D-Value tracking was fixed. The -eo flag currently causes
infinite recursions on a test case where -ew or -ek do not. The cause is
unknown.